app.py
import os
import torch
import numpy as np
from flask import Flask, request, jsonify
from dotenv import load_dotenv
from PIL import Image
import io
from waitress import serve

# 1. Secrets Injection
load_dotenv()
API_KEY = os.getenv("MY_API_SECRET") # Example of a managed secret

app = Flask(__name__)

# 2. Model Replacement (Loading your trained segmentation model)
# Ensure your model file is in the same directory or mapped via Docker
MODEL_PATH = "house_segmentation_model.pth"
device = torch.device("cuda" if torch.cuda.is_available() else "cpu")

# Placeholder for a U-Net or similar architecture
# model = YourModelClass() 
# model.load_state_dict(torch.load(MODEL_PATH, map_location=device))
# model.eval()

@app.route('/predict', methods=['POST'])
def predict():
    if 'image' not in request.files:
        return jsonify({"error": "No image provided"}), 400
    
    file = request.files['image']
    img_bytes = file.read()
    image = Image.open(io.BytesIO(img_bytes)).convert("RGB")
    
    # Preprocessing (Match your training preprocessing)
    # 1. Resize, 2. ToTensor, 3. Normalize
    
    # Inference logic
    # with torch.no_grad():
    #     prediction = model(input_tensor)
    
    return jsonify({"status": "success", "message": "Mask generated (placeholder)"})

if __name__ == '__main__':
    print("Segmentation Server starting on port 5000...")
    serve(app, host='0.0.0.0', port=5000)